Patches by Andreas Steinmetz: 00-lib-poll.patch
	- select() durch poll() ersetzen

01-lib-performance.patch

	- unnötige Divisionen und Multiplikationen ersetzen
	- Verwendung einer verketteten Liste anstelle
	  von kompletten Array-Prüfungen

02-lib-warnings.patch

	- Compiler Warnungen beseitigen

03-common-warnings.patch

	- Compiler Warnungen beseitigen

04-netcv2dvbip-warnings.patch

	- Compiler Warnungen beseitigen

05-netcv2dvbip-options.patch

	- neue Option -q für weniger Bildschirmausgaben
	- neue Option -o <directory> für Ausgabeverzeichnis
	  für channels.m3u

06-netcv2dvbip-maxapids.patch

	- Anpassung von MAXAPIDS an die Realität

07-netcv2dvbip-poll.patch

	- select() durch poll() ersetzen

08-netcv2dvbip-mutex-cleanup.patch

	- vorbereitendes Aufräumen der Mutexes
	- Verschieben einer Definition

09-netcv2dvbip-buffers.patch

	- Verwendung von verketteten Pufferelementen
	  und eines Pools freier Elemente anstatt
	  eines Round Robin Puffers
	- minimierte Lockingzeiten der Haupt-Mutex

10-netcv2dvbip-nocopy.patch

	- IPv4 Multicast Versand ohne zusätzliches
	  Umkopieren der Daten
	- minimierte Lockingzeiten der Haupt-Mutex

11-all-oomhandling.patch

	- korrekte Behandlung von Speicherallokierungsfehlern

12-netcv2dvbip-linux.patch

	- die Option -t steht nur auf geeigneten Linux-Systemen
	  zur Verfügung (Compile-Time Lösung)

13-all-mingw32.patch

	- funktionierender Build von mcli, netcv2dvbip, netcvdiag
	  und netcvlogview mit MingGW unter WIN32, ich habe die
	  Tools kurz angetestet und sie scheinen zu funktionieren,
	  beinhaltet auch einen WIN32-Ersatz für poll() - mcli
	  funktioniert zumindest zur Verwendung von netcvdiag
	  hinreichend

	  Hinweis: In client, netcv2dvbip und tool gibt es jetzt
	  jeweils ein Unterverzeichnis "mingw". In dem jeweiligen
	  mingw Verzeichnis build.cmd aufrufen (ggf. muss in dem
	  Batch File der Pfad zu MinGW angepasst werden).
	  netcvupdate habe ich bewusst nicht portiert, zum einen
	  ist hier der Aufwand höher, zum anderen fehlen die
	  aufgerufenen POSIX-Tools. Ausserdem will ich mir meinen
	  Netceiver nicht mit Tests ruinieren. Den brauche ich zum
	  Fernsehen 

14-lib-symlink.tar

	- benötigter Symlink für 13-all-mingw32.patch
	  mcast/client/poll.c muss ein Symlink auf
	  mcast/common/win32/poll.c sein

+++ mcast/common/recv_ccpp.c
@@ -256,7 +256,7 @@ int get_tra_data (xmlChar * xmlbuff, int buffersize, tra_info_t * tra_info)
 							if (c.key) {
 								dbg ("SymbolRate: %s\n", c.key);
 								int val=atoi ((char *) c.key);
-								switch(tra->fe_type) {
+								switch((int)tra->fe_type) {
 									case FE_DVBS2:
 									case FE_QPSK:
 										tra->fep.u.qpsk.symbol_rate=val;
@@ -275,7 +275,7 @@ int get_tra_data (xmlChar * xmlbuff, int buffersize, tra_info_t * tra_info)
 							if (c.key) {
 								dbg ("FecInner: %s\n", c.key);
 								int val=atoi ((char *) c.key);
-								switch(tra->fe_type) {
+								switch((int)tra->fe_type) {
 									case FE_DVBS2:
 									case FE_QPSK:
 										tra->fep.u.qpsk.fec_inner=(fe_code_rate_t)val;
@@ -294,7 +294,7 @@ int get_tra_data (xmlChar * xmlbuff, int buffersize, tra_info_t * tra_info)
 							if (c.key) {
 								dbg ("Modulation: %s\n", c.key);
 								int val=atoi ((char *) c.key);
-								switch(tra->fe_type) {
+								switch((int)tra->fe_type) {
 									case FE_QAM:
 										tra->fep.u.qam.modulation=(fe_modulation_t)val;
 										break;
